requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
flask-talisman==1.1.0
